move review tool in to /app
